Avoid problems with Kiwi Syslog Server in Varian TrueBeam 4.0 project
Adapt to Kiwi Syslog server and configuration interface not working anymore due to crash: Implement a server that works with the SysLog implementation in the Varian TrueBeam 4.0 QA Data Interface project.
For viewing the logs, VSCode works well because it updates the view after file changes and colorizes the text for better readability.